requirements.txt
# Existing dependencies
requests
pyyaml
openai
pydantic

# NEW - Research Intelligence System Dependencies

# Layer 1: Deep Paper Analysis
google-genai>=0.1.0         # NEW Gemini SDK with structured output support
PyMuPDF>=1.24.0             # PDF text extraction (fitz)

# Optional: Alternative LLM providers
# anthropic>=0.40.0         # Claude API (can configure in src/config.py)

# Layer 2: Bibliometric Analysis
networkx>=3.2             # Citation graphs and network analysis
python-louvain>=0.16      # Community detection for research fronts

# Layer 3: Visualization & Email
matplotlib>=3.8.0         # Graph visualization
seaborn>=0.13.0          # Enhanced plotting
# Email delivery uses smtplib (stdlib) via Gmail — no extra package needed
